refactor(frontend): break down InterviewPrep.jsx into feature folder structure
